requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
motor==3.1.1
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
google-generativeai